[#IOPID-2345] Rename Activity
